Renamed buttons to fix typo, and removed option for undefined onclick. Changed props for SegmentedButtons so they only take one list
